Added STL files.
Added ShellBottom.stl and ShellTop.stl

These are STL files designed as a casing to contain the power supply, Arduino, buck converter, and all other associated components in one compact package. There are also cutouts designed for a power inlet and an overall power switch that are accessible to the user. The case largely serves as an aesthetic value, cleaning up all the loose components, however it also serves a function which is safely covering live electrical terminals on the power supply.